Fix first_level_features() import and move dataset.py out of preprocessing (and update all scripts)
