+++ CICIDS2017/dataset.py
@@ -19,7 +19,7 @@ from CICIDS2017.preprocessing.subset import subset_indices
 from CICIDS2017.analysis.distribution import data_distribution
 from CICIDS2017.analysis.mutual_info import mutual_info_classif
 from CICIDS2017.analysis.pca import apply_pca
-from CICIDS2017.analysis.first_level_features import first_level_features
+from CICIDS2017.analysis.features import first_level_features
 
 class CICIDS2017:
     def __init__(self, dataset_size=None, logger=SimpleLogger()):
