+++ NIDS_attacks/decision_tree_attack/hsj_attack_decision_tree.py
@@ -8,8 +8,8 @@ sys.path.append(root_dir)
 
 import numpy as np
 
-from CICIDS2017.preprocessing.dataset import CICIDS2017
-from UNSWNB15.preprocessing.dataset import UNSWNB15
+from CICIDS2017.dataset import CICIDS2017
+from UNSWNB15.dataset import UNSWNB15
 
 from scripts.models.decision_tree.decision_tree import train_decision_tree
 from scripts.logger import LoggerManager
